Move tests inside integration module
- Remove the Example suffix from all tests
- Move tests to snunit.test package
- Update Scalafmt to 2.7.2
