+++ .scalafmt.conf
@@ -1,2 +1,2 @@
-version = "2.6.1"
+version = "2.7.2"
 maxColumn = 120
