--- examples/multiple-handlers/src/snunit/examples/MultipleHandlersExample.scala
@@ -1,29 +1,0 @@
-package snunit.examples
-
-import snunit._
-
-object MultipleHandlersExample {
-  def main(args: Array[String]): Unit = {
-    val server =
-      SyncServerBuilder()
-        .withRequestHandler(req => {
-          if (req.method == Method.GET) {
-            req.send(
-              statusCode = 200,
-              content = s"Hello world multiple handlers!\n",
-              headers = Seq("Content-Type" -> "text/plain")
-            )
-          } else req.next()
-        })
-        .withRequestHandler(req => {
-          req.send(
-            statusCode = 404,
-            content = s"Not found\n",
-            headers = Seq("Content-Type" -> "text/plain")
-          )
-        })
-        .build()
-
-    server.listen()
-  }
-}
